+++ changes.md
@@ -1,5 +1,8 @@
 # Zealot
 
+## Release v0.3.0
+LTS: 29/09/2025
+
  - Using the result of `HttpInterceptor.onFollow` when performing redirect requests 
 
 ## Release v0.2.9